Create a global settings file for the CLI
